% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calculations.R
\name{bedd}
\alias{bedd}
\title{Calculate Biologically Effective Growing Days value (BEDD)}
\usage{
bedd(data)
}
\arguments{
\item{data}{data frame or tibble as downloaded with \code{\link[weathercan]{weather_dl}}. Provide a single measurement site for a single year only.}
}
\value{
a named list including the Biologically Effective Degere Days and binned 'Region'. See Gladstones (1992)
}
\description{
Calculate Biologically Effective Growing Days value (BEDD)
}
