% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calculations.R
\name{calculate_annual_indicies}
\alias{calculate_annual_indicies}
\title{Calculate Annual Temperature Indices}
\usage{
calculate_annual_indicies(data)
}
\arguments{
\item{data}{data frame or tibble as downloaded with \code{\link[weathercan]{weather_dl}}. Multiple stations or years are accepted. Daily data required.}
}
\value{
a tibble with climate indicies and region classification for each climate station for each year.
}
\description{
Calculate GST, WI, HI, and BEDD values for each year of data provided, at each individual climate station (separated by ID).
}
