% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calculations.R
\name{frost}
\alias{frost}
\title{Frost Free Length & Frosty Days}
\usage{
frost(data)
}
\arguments{
\item{data}{data frame or tibble as downloaded with \code{\link[weathercan]{weather_dl}}. Provide a single measurement site for a single year only.}
}
\value{
The longest number of days between dates with min_temp < 0 C, and number of Frosty Days
}
\description{
Frost Free Length & Frosty Days
}
