% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calculations.R
\name{get_region}
\alias{get_region}
\title{Get Region classification}
\usage{
get_region(calculation = "WI", value)
}
\arguments{
\item{calculation}{Which calculation's value of which to look up the region result.}

\item{value}{The value to look up}
}
\value{
The region (from Too Cool to Too Hot)
}
\description{
Get Region classification
}
