% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrapeWeatheR-pkg.R
\docType{package}
\name{GrapeWeatheR-package}
\alias{GrapeWeatheR-package}
\alias{GrapeWeatheR}
\title{GrapeWeatheR}
\description{
Analysis of Environment Canada Climate Data for Grape Grower and Agriculture Information
}
\references{
Gladstones, J. 1992. Viticulture and Environment. Winetitles, Adelaide.
Huglin, P. 1978. Nouveau Mode d'Evaluation des Possibilites Heliothermiques d'un Milieu Viticole. C.R. Acad. Agr. France 64:1117-1126
Winkler, A.J., J.A. Cook, W.M. Kliewer, and L.A. Lider. 1974. General Viticulture. 4th ed. University of California Press, Berkeley.
}
