% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calculations.R
\name{gst}
\alias{gst}
\title{Calculate Growing Season Temperature average, max and min (GST)}
\usage{
gst(data)
}
\arguments{
\item{data}{data frame or tibble as downloaded with \code{\link[weathercan]{weather_dl}}. Provide a single measurement site for a single year only.}
}
\value{
named list of GST average temperature, GST max temperature and GST min temperature, plus GST binned 'Region'
}
\description{
Calculate Growing Season Temperature average, max and min (GST)
}
