% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calculations.R
\name{huglin_index}
\alias{huglin_index}
\title{Calculate Huglin Index (HI)}
\usage{
huglin_index(data)
}
\arguments{
\item{data}{data frame or tibble as downloaded with \code{\link[weathercan]{weather_dl}}. Provide a single measurement site for a single year only.}
}
\value{
a named list including the Huglin Index value and binned 'Region'. See Huglin (1978)
}
\description{
Calculate Huglin Index (HI)
}
