% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calculations.R
\name{lat_correction}
\alias{lat_correction}
\title{Calculate Latitude Correction Values for HI and BEDD. HI = 1.02 for 40, 1.06 for 50. Calculated for Apr 01-Sept 31}
\usage{
lat_correction(lat, ...)
}
\arguments{
\item{lat}{latitude to calculate K}

\item{...}{additional hidden calculation parameters}
}
\value{
named list of HI and BEDD K values for the specified latitude
}
\description{
Calculate Latitude Correction Values for HI and BEDD. HI = 1.02 for 40, 1.06 for 50. Calculated for Apr 01-Sept 31
}
