% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acquisition.R
\name{load_weather_station_data}
\alias{load_weather_station_data}
\title{Use the weathercan package to download data}
\usage{
load_weather_station_data(station_list, ...)
}
\arguments{
\item{station_list}{a single station_id or a vector of stations to download from. Stations can be found using \code{\link[weathercan]{stations_search}}.}

\item{...}{additional parameters from weathercan. See \code{\link[weathercan]{weather_dl}} for more information.}
}
\value{
the downloaded weather data.
}
\description{
Use the weathercan package to download data
}
