% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports_and_plots.R
\name{plot_index_accumilation}
\alias{plot_index_accumilation}
\title{Plot index accumilation}
\usage{
plot_index_accumilation(data, index = "Winkler", year = NULL)
}
\arguments{
\item{data}{Raw hourly or daily data returned from weathercan}

\item{index}{The index of which to plot accumilation (one of 'Winkler', 'Huglin', or 'BEDD')}

\item{year}{year to plot, optional (or leave NULL for all)}
}
\value{
an annual accumilation plot of the specified index
}
\description{
Plot index accumilation
}
