% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports_and_plots.R
\name{plot_index_history}
\alias{plot_index_history}
\title{Plot the annual index value over time}
\usage{
plot_index_history(data, index = "Winkler")
}
\arguments{
\item{data}{Data frame wfrom which to plot}

\item{index}{The index of which to plot annual values (one of 'WI', 'HI', 'BEDD', or 'GST', default 'WI')}
}
\value{
a plot
}
\description{
Plot the annual index value over time
}
