% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports_and_plots.R
\name{plot_precip_history}
\alias{plot_precip_history}
\title{Plot Precipitation History}
\usage{
plot_precip_history(data, interval = "day", trim_years = 5)
}
\arguments{
\item{data}{Data frame from which to plot.}

\item{interval}{Whether to plot `day` or `week` or `month` or `year` amounts}

\item{trim_years}{the number of years of data to plot.}
}
\value{
a ggplot object
}
\description{
Plot Precipitation History
}
