% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports_and_plots.R
\name{plot_precip_progress}
\alias{plot_precip_progress}
\title{Plot Precipitation Progress}
\usage{
plot_precip_progress(data, year = NULL)
}
\arguments{
\item{data}{Data from which to plot}

\item{year}{year to plot, optional (NULL to plot all)}
}
\value{
a ggplot plot
}
\description{
Plot Precipitation Progress
}
