% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports_and_plots.R
\name{plot_temp_history}
\alias{plot_temp_history}
\title{Plot Annual Temperature History}
\usage{
plot_temp_history(data)
}
\arguments{
\item{data}{Data frame from which to plot.}
}
\value{
a ggplot object
}
\description{
Plot Annual Temperature History
}
