% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calculations.R
\name{precip}
\alias{precip}
\title{Calculate Precipitation Statistics}
\usage{
precip(data)
}
\arguments{
\item{data}{data frame or tibble as downloaded with \code{\link[weathercan]{weather_dl}}. Provide a single measurement site for a single year only.}
}
\value{
precipitation values for the site.
}
\description{
Calculate Precipitation Statistics
}
