% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calculations.R
\name{turn_hour_data_to_daily}
\alias{turn_hour_data_to_daily}
\title{Turn hourly data to daily data}
\usage{
turn_hour_data_to_daily(data)
}
\arguments{
\item{data}{data frame or tibble as downloaded with \code{\link[weathercan]{weather_dl}}. Passes non-hourly data back without modification.}
}
\value{
Data turned into daily data, or returned data otherwise.
}
\description{
Turn hourly data to daily data
}
