% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calculations.R
\name{winkler_index}
\alias{winkler_index}
\title{Calculate Winkler Index (Growing Degree Days) (WI/GDD)}
\usage{
winkler_index(data)
}
\arguments{
\item{data}{data frame or tibble as downloaded with \code{\link[weathercan]{weather_dl}}. Provide a single measurement site for a single year only.}
}
\value{
a named list including the Winkler Index value and binned 'Region'. See Winkler et. al. (1974)
}
\description{
Calculate Winkler Index (Growing Degree Days) (WI/GDD)
}
