% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports_and_plots.R
\name{plot_precip_history}
\alias{plot_precip_history}
\title{Plot Precipitation History}
\usage{
plot_precip_history(data)
}
\arguments{
\item{data}{Data frame from which to plot.}
}
\value{
a ggplot object
}
\description{
Plot Precipitation History
}
