% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports_and_plots.R
\name{plot_temp_history}
\alias{plot_temp_history}
\title{Plot Annual Temperature History}
\usage{
plot_temp_history(data, average_n = 1, interval = "day", what = "max")
}
\arguments{
\item{data}{Data frame from which to plot.}

\item{average_n}{The number of items to include in a rolling average}

\item{interval}{One of 'day', 'week', 'month'}

\item{what}{One of 'max', 'mean', 'min'}
}
\value{
a plot
}
\description{
Plot Annual Temperature History
}
